functions/requirements.txt
cachecontrol==0.12.12; python_version >= "3.6"
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
certifi==2022.9.14; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy"
cffi==1.15.1
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy" and python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7" and python_version < "4"
cloudevents==1.6.1
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and python_version < "4" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0; python_version >= "3.5" and python_version < "4"
firebase-admin==5.3.0; python_version >= "3.6"
flask==2.0.2; python_version >= "3.6"
functions-framework==3.0.0; python_version >= "3.5" and python_version < "4"
google-api-core==2.10.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-api-python-client==2.61.0; python_version >= "3.7"
google-auth-httplib2==0.1.0; python_version >= "3.7"
google-auth==2.11.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
google-cloud-core==2.3.2; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-cloud-firestore==2.6.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-cloud-pubsub==2.13.6; python_version >= "3.7"
google-cloud-storage==2.5.0; python_version >= "3.7"
google-crc32c==1.5.0; python_version >= "3.7"
google-resumable-media==2.3.3; python_version >= "3.7"
googleapis-common-protos==1.56.4; platform_python_implementation != "PyPy" and python_version >= "3.7"
grpc-google-iam-v1==0.12.4; python_version >= "3.7"
grpcio-status==1.49.0; python_version >= "3.7" and platform_python_implementation != "PyPy"
grpcio==1.49.0; python_version >= "3.7" and platform_python_implementation != "PyPy"
gunicorn==20.1.0; python_version >= "3.5" and python_version < "4" and platform_system != "Windows"
httplib2==0.20.4; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
idna==3.4; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy"
itsdangerous==2.1.2; python_version >= "3.7" and python_version < "4"
jinja2==3.1.2; python_version >= "3.7" and python_version < "4"
loguru==0.6.0; python_version >= "3.5"
markupsafe==2.1.1; python_version >= "3.7" and python_version < "4"
msgpack==1.0.4; python_version >= "3.6"
numpy==1.23.3; python_version >= "3.8"
packaging==21.3; python_version >= "3.6" and python_version < "4"
proto-plus==1.22.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
protobuf==4.21.6; platform_python_implementation != "PyPy" and python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy") or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
pycparser==2.21; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pympi-ling==1.70.2
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7" and python_version < "4"
requests==2.28.1; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy"
rsa==4.9; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
soundfile==0.10.3.post1
uritemplate==4.1.1; python_version >= "3.7"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" and platform_python_implementation != "PyPy"
watchdog==1.0.2; python_version >= "3.6" and python_version < "4"
werkzeug==2.2.2; python_version >= "3.7" and python_version < "4"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
